+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Use `jni` and `jnigen` instead of method channels
     * Due to this more methods are now synchronous
 * Add more properties to `PackageInfo`
-* Add `getAppIcon` method
+* Add `getIcon` and `getIconFromApk` method
 
 ## 0.0.1